feat(nes): update nes emulator to use hal video task for video output instead of old nes-specific video task
